+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ## v0.11.6
 * added new tokens to Market APIs
   * added "last week" time options
+* moved `BTCSearchBox` to `gui.customswing`
 
 ## v0.11.5
 * added character limits matching database field sizes